[coinbaseex] move test to Integration class
