+++ xchange-coinbaseex/src/test/java/com/xeiam/xchange/coinbaseex/CoinbaseExExchangeTest.java
@@ -15,15 +15,4 @@ public class CoinbaseExExchangeTest {
   public void coinbaseShouldBeInstantiatedWithoutAnExceptionWhenUsingDefaultSpecification() {
     ExchangeFactory.INSTANCE.createExchange(CoinbaseExExchange.class.getCanonicalName());
   }
-
-  @Test
-  public void testCoinbaseCurrencies() throws IOException {
-    CoinbaseExExchange exchange = new CoinbaseExExchange();
-    exchange.applySpecification(exchange.getDefaultExchangeSpecification());
-
-    assertThat(exchange.getPollingMarketDataService().getExchangeSymbols()).containsAll(
-        Arrays.asList(new CurrencyPair("BTC/USD"), new CurrencyPair("BTC/GBP"), new CurrencyPair("BTC/EUR"))
-    );
-  }
-
 }
